Finish evaluation information

+++ Code/Client/app/src/main/java/com/example/cz17a/quizclient/Src/Jackpot.java
@@ -48,7 +48,7 @@ public class Jackpot {
         this.isActive = false;
     }
 
-    public int payout(int responseTime){
+    public int payout(double responseTime){
         if(isActive) {
             if (responseTime < 5) {
                 return this.amount;
